Bump Java source/target from 1.7 to 1.8
Although the code of this project likely is compatible with Java 1.4, modern JVMs refuse to compile, with errors like these:

```
Compilation failure: Compilation failure:
Error:  Source option 7 is no longer supported. Use 8 or later.
Error:  Target option 7 is no longer supported. Use 8 or later.
```

To circument that, the source and darget definitions are set to 8 in this commit.